Refactor apps
