+++ course_discovery/apps/api/tests/test_serializers.py
@@ -2,8 +2,7 @@ from django.core.urlresolvers import reverse
 from django.test import TestCase, RequestFactory
 
 from course_discovery.apps.api.serializers import CatalogSerializer, CourseSerializer, ContainedCoursesSerializer
-from course_discovery.apps.catalogs.tests.factories import CatalogFactory
-from course_discovery.apps.courses.tests.factories import CourseFactory
+from course_discovery.apps.core.tests.factories import CatalogFactory, CourseFactory
 
 
 class CatalogSerializerTests(TestCase):
