+++ course_discovery/apps/api/v1/tests/test_views.py
@@ -12,11 +12,10 @@ from rest_framework.reverse import reverse
 from rest_framework.test import APITestCase, APIRequestFactory
 
 from course_discovery.apps.api.serializers import CatalogSerializer, CourseSerializer
-from course_discovery.apps.catalogs.models import Catalog
-from course_discovery.apps.catalogs.tests.factories import CatalogFactory
-from course_discovery.apps.core.tests.factories import UserFactory, USER_PASSWORD
-from course_discovery.apps.core.tests.mixins import ElasticsearchTestMixin
-from course_discovery.apps.courses.tests.factories import CourseFactory
+from course_discovery.apps.base.tests.factories import UserFactory, USER_PASSWORD
+from course_discovery.apps.base.tests.mixins import ElasticsearchTestMixin
+from course_discovery.apps.core.models import Catalog
+from course_discovery.apps.core.tests.factories import CatalogFactory, CourseFactory
 
 OAUTH2_ACCESS_TOKEN_URL = 'http://example.com/oauth2/access_token/'
