+++ course_discovery/apps/base/constants.py
@@ -1,0 +1,7 @@
+""" Constants for the base app. """
+
+
+class Status(object):
+    """Health statuses."""
+    OK = u"OK"
+    UNAVAILABLE = u"UNAVAILABLE"
