+++ course_discovery/apps/base/tests/test_context_processors.py
@@ -2,13 +2,13 @@
 
 from django.test import TestCase, override_settings, RequestFactory
 
-from course_discovery.apps.core.context_processors import core
+from course_discovery.apps.base.context_processors import core
 
 PLATFORM_NAME = 'Test Platform'
 
 
 class CoreContextProcessorTests(TestCase):
-    """ Tests for core.context_processors.core """
+    """ Tests for base.context_processors.core """
 
     @override_settings(PLATFORM_NAME=PLATFORM_NAME)
     def test_core(self):
