+++ course_discovery/apps/core/admin.py
@@ -1,30 +1,12 @@
-""" Admin configuration for core models. """
-
 from django.contrib import admin
-from django.contrib.auth.admin import UserAdmin
-from django.utils.translation import ugettext_lazy as _
 
-from course_discovery.apps.core.forms import UserThrottleRateForm
-from course_discovery.apps.core.models import User, UserThrottleRate
+from course_discovery.apps.core.models import Catalog
 
 
-class CustomUserAdmin(UserAdmin):
-    """ Admin configuration for the custom User model. """
-    list_display = ('username', 'email', 'full_name', 'first_name', 'last_name', 'is_staff')
-    fieldsets = (
-        (None, {'fields': ('username', 'password')}),
-        (_('Personal info'), {'fields': ('full_name', 'first_name', 'last_name', 'email')}),
-        (_('Permissions'), {'fields': ('is_active', 'is_staff', 'is_superuser',
-                                       'groups', 'user_permissions')}),
-        (_('Important dates'), {'fields': ('last_login', 'date_joined')}),
-    )
+@admin.register(Catalog)
+class CatalogAdmin(admin.ModelAdmin):
+    list_display = ('name',)
+    readonly_fields = ('created', 'modified',)
 
-
-class UserThrottleRateAdmin(admin.ModelAdmin):
-    """ Admin configuration for the UserThrottleRate model. """
-    form = UserThrottleRateForm
-    raw_id_fields = ('user',)
-
-
-admin.site.register(User, CustomUserAdmin)
-admin.site.register(UserThrottleRate, UserThrottleRateAdmin)
+    class Media(object):
+        js = ('js/catalogs-change-form.js',)
