+++ course_discovery/apps/core/exceptions.py
