+++ course_discovery/apps/core/tests/factories.py
@@ -1,16 +1,34 @@
 import factory
+from factory.fuzzy import FuzzyText
 
-from course_discovery.apps.core.models import User
-
-USER_PASSWORD = 'password'
+from course_discovery.apps.core.models import Catalog, Course
 
 
-class UserFactory(factory.DjangoModelFactory):
-    username = factory.Sequence(lambda n: 'user_%d' % n)
-    password = factory.PostGenerationMethodCall('set_password', USER_PASSWORD)
-    is_active = True
-    is_superuser = False
-    is_staff = False
+class CatalogFactory(factory.DjangoModelFactory):
+    class Meta(object):
+        model = Catalog
 
-    class Meta:
-        model = User
+    name = FuzzyText(prefix='catalog-name-')
+    query = '{"query": {"match_all": {}}}'
+
+
+class CourseFactory(factory.Factory):
+    class Meta(object):
+        model = Course
+        exclude = ('name',)
+
+    id = FuzzyText(prefix='course-id/', suffix='/fake')
+    name = FuzzyText(prefix="էҽʂէ çօմɾʂҽ ")
+
+    @factory.lazy_attribute
+    def body(self):
+        return {
+            'id': self.id,
+            'name': self.name
+        }
+
+    @classmethod
+    def _create(cls, model_class, *args, **kwargs):
+        obj = model_class(*args, **kwargs)
+        obj.save()
+        return obj
