+++ course_discovery/apps/core/tests/test_refresh_all_courses.py
@@ -6,7 +6,7 @@ from django.test.utils import override_settings
 from edx_rest_api_client.client import EdxRestApiClient
 from mock import patch
 
-from course_discovery.apps.courses.models import Course
+from course_discovery.apps.core.models import Course
 
 
 @override_settings(
