--- course_discovery/apps/courses/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class CoursesConfig(AppConfig):
-    name = 'courses'
-    verbose_name = 'Courses'
