--- course_discovery/apps/courses/constants.py
@@ -1,2 +1,0 @@
-COURSE_ID_REGEX = r'[^/+]+(/|\+)[^/+]+(/|\+)[^/]+'
-COURSE_ID_PATTERN = r'(?P<id>{})'.format(COURSE_ID_REGEX)
