--- course_discovery/apps/courses/views.py
@@ -1,5 +1,0 @@
-from django.views.generic import TemplateView
-
-
-class QueryPreviewView(TemplateView):
-    template_name = 'catalogs/preview.html'
